# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: FameGame API
  version: "1.0.0"
  license:
    name: Proprietary
  description: |
    Product HTTP contracts for FameGameBackend.

    Success payloads use an envelope:
    `{ "data": <payload>, "meta": { "message": string|null } }`.

    Errors use RFC 7807 `application/problem+json`.
    Most `/api/*` routes require JWT (`Authorization: Bearer <token>`).
    Public auth routes set `security: []`.
servers:
  - url: /api
    description: API prefix
...
security:
  - bearerAuth: []
paths:
...
  /game-shop/equip:
    post:
      operationId: gameShopEquip
      tags: [GameShop]
      summary: Equip inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemId]
              properties:
                itemId:
                  type: integer
      responses:
        200:
          description: Equipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        401:
          description: Unauthenticated / invalid JWT
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
        403:
          description: Forbidden / business rule
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
  /game-shop/purchase:
    post:
      operationId: gameShopPurchase
      tags: [GameShop]
      summary: Buy item from shop slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [storeSlotId]
              properties:
                storeSlotId:
                  type: integer
                chestSlotIndex:
                  type: integer
      responses:
        200:
          description: Purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        401:
          description: Unauthenticated / invalid JWT
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
        403:
          description: Forbidden / business rule
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
  /game-shop/refresh:
    post:
      operationId: gameShopRefresh
      tags: [GameShop]
      summary: Refresh shop offers
      responses:
        200:
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        401:
          description: Unauthenticated / invalid JWT
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
        403:
          description: Forbidden / business rule
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
  /game-shop/sell:
    post:
      operationId: gameShopSell
      tags: [GameShop]
      summary: Sell inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [storageSlotId]
              properties:
                storageSlotId:
                  type: integer
      responses:
        200:
          description: Sold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        401:
          description: Unauthenticated / invalid JWT
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
        403:
          description: Forbidden / business rule
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
  /game-shop/sell-equipped:
    post:
      operationId: gameShopSellEquipped
      tags: [GameShop]
      summary: Sell currently equipped item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slotType]
              properties:
                slotType:
                  type: string
                  enum: [helmet, weapon, armor, shield, gloves, boots, amulet, ring, potions]
      responses:
        200:
          description: Sold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        401:
          description: Unauthenticated / invalid JWT
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
        403:
          description: Forbidden / business rule
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
  /game-shop/state:
    get:
      operationId: gameShopState
      tags: [GameShop]
      summary: "Full shop, inventory and equipped state"
      responses:
        200:
          description: State
          content:
            application/json:
              schema:
                allOf:
                  -
                    $ref: "#/components/schemas/ApiEnvelope"
                  -
                    type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/GameShopState"
        401:
          description: Unauthenticated / invalid JWT
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
        403:
          description: Forbidden / business rule
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
...
  /user_equipments/{id}/equip:
    post:
      operationId: equipItem
      tags: [Equipment]
      summary: Equip item onto equipment resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemId]
              properties:
                itemId:
                  type: integer
      responses:
        200:
          description: Equipped
          content:
            application/json:
              schema:
                allOf:
                  -
                    $ref: "#/components/schemas/ApiEnvelope"
                  -
                    type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/EquippedResult"
        401:
          description: Unauthenticated / invalid JWT
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
        403:
          description: Forbidden / business rule
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemJson"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          description: Endpoint-specific success payload
          additionalProperties: true
        meta:
          type: object
          properties:
            message:
              type: ["string", "null"]
    ProblemJson:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        violations:
          type: array
          items:
            type: object
            properties:
              propertyPath:
                type: string
              message:
                type: string
...
    EquippedResult:
      type: object
      properties:
        equipped:
          type: boolean
...
    GameShopRefresh:
      type: object
      properties:
        isFreeRefreshAvailable:
          type: boolean
        refreshCost:
          type: integer
    GameShopState:
      type: object
      properties:
        gold:
          type: integer
        shop:
          type: array
          items:
            type: ["object", "null"]
        inventory:
          type: array
          items:
            type: ["object", "null"]
        equipped:
          type: object
          additionalProperties:
            type: ["object", "null"]
        refresh:
          $ref: "#/components/schemas/GameShopRefresh"
...